+++ scripts/basic_checks.sh
@@ -1,2 +1,3 @@
 skip-trace --help
 skip-trace who-owns matthew
+skip-trace explain matthew