+++ scripts/make_source.sh
@@ -1,3 +1,8 @@
 git2md skip_trace \
   --ignore __pycache__ \
   --output SOURCE.md
+
+
+git2md /C/github/skip_trace/.venv/Lib/site-packages/sigstore \
+  --ignore __pycache__ \
+  --output SOURCE_SIGSTORE.md