+++ skip_trace/__about__.py
@@ -5,15 +5,25 @@ __all__ = [
     "__version__",
     "__description__",
     "__readme__",
+    "__license__",
     "__credits__",
     "__requires_python__",
     "__status__",
+    "__keywords__",
 ]
 
 __title__ = "skip-trace"
-__version__ = "0.1.0"
+__version__ = "0.1.1"
 __description__ = "Ownership Attribution for Python Packages"
 __readme__ = "README.md"
+__license__ = "MIT"
 __credits__ = [{"name": "Matthew Dean Martin", "email": "matthewdeanmartin@gmail.com"}]
 __requires_python__ = ">=3.8"
 __status__ = "1 - Planning"
+__keywords__ = [
+    "PyPI maintainers",
+    "package owners",
+    "package provenance",
+    "software supply chain",
+    "PEP 541",
+]