+++ skip_trace/__init__.py
@@ -1,6 +1,4 @@
 # skip_trace/__init__.py
 
-__version__ = "0.1.0"
-
 # __all__ will be populated as public functions/classes are added.
 __all__ = []