+++ CHANGELOG.md
@@ -12,6 +12,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed for any bug fixes.
 - Security in case of vulnerabilities.
 
+## [0.1.1] - 2025-10-12
+
+### Added
+
+- PKG_INFO collector
+- Scans URLs/home pages for contact info
+
 ## [0.1.0] - 2025-10-12
 
 ### Added